minor: navigated image rendering to decktape method and added test(wip)
