+++ test/modules/asciidoc-renderer/pdf-renderer.test.ts
@@ -1,27 +1,33 @@
-import {AsciidocCompiler, AsciidocFile} from "@solardoc/asciidoc-renderer";
-import {PDFRenderer} from "@solardoc/asciidoc-renderer/dist/renderer/targets/pdf";
-import {assert} from "chai";
+import {AsciidocCompiler, AsciidocFile, DecktapeSlim, HTMLRenderer, PDFRenderer} from "@solardoc/asciidoc-renderer";
+import { assert } from "chai";
+import path from "path";
+import fs from "fs";
 
-describe("PDFRenderer", () => {
-  describe("NodePDFRenderer", () => {
-    describe("render()", () => {
+
+describe("PdfRenderer", () => {
+  describe("render()", () => {
+    it("should return PDF document", async () => {
       const adocString =
-        `= Test\\n\\nx\\n\\n== Still testing\\n\\nMain-Slide 2\\n\\n=== Still testing\\n\\nSub-Slide 2.1\\n\\n== Still testing
-\\n\\nx\\n\\n== Still testing\\n\\nx"`;
+        `= Test\n\nx\n\n== Still testing\n\nMain-Slide 2\n\n=== Still testing\n\nSub-Slide 2.1\n\n== Still testing
+\n\nx\n\n== Still testing\n\nx"`;
       const adocFilename = "test.adoc";
+      const adocFile = await AsciidocFile.fromString(
+        adocFilename,
+        adocString,
+      );
+      const asciidocCompiler = new AsciidocCompiler();
+      const presentation = await asciidocCompiler.parse(adocFile);
 
-      // TODO! Implement proper PDFRenderer tests once the PDFRenderer is implemented
-      // it("should return a PDFOutput", async () => {
-      //   const adocFile = await AsciidocFile.fromString(
-      //     adocString,
-      //     adocFilename
-      //   );
-      //   const asciidocCompiler = new AsciidocCompiler();
-      //   const presentation = await asciidocCompiler.compile(adocFile);
-      //
-      //   const html = await presentation.render(new PDFRenderer());
-      //   assert.notEqual(html.internalData, "", "internalData is empty");
-      // });
+      const pdf = await presentation.render(new PDFRenderer(),{revealJSAssetsPath: "https://cdn.jsdelivr.net/npm/reveal.js@5.0.2/"});
+      assert.notEqual(pdf.internalData, null, "internalData is empty");
+      const pdfDir = path.dirname("test.pdf");
+      try {
+        fs.accessSync(pdfDir, fs.constants.F_OK);
+      } catch {
+        fs.mkdirSync(pdfDir, { recursive: true });
+      }
+      fs.writeFileSync("test.pdf", await pdf.internalData.save({ addDefaultPage: false }));
+
     });
-  })
+  });
 });
